refactor: move SkillsBench benchmark helpers into adapter
