+++ goal_harness/cli.py
@@ -19,9 +19,17 @@ from .bootstrap import (
     render_bootstrap_markdown,
 )
 from .benchmark import (
-    AGENTS_LAST_EXAM_DEFAULT_ALT_DOCKER_IMAGE,
-    AGENTS_LAST_EXAM_DEFAULT_DOCKER_IMAGE,
-    AGENTS_LAST_EXAM_DEFAULT_SNAPSHOT,
+    benchmark_result_from_benchmark_run_for_baseline_gate,
+    build_benchmark_adapter_kwarg_absorption_review,
+    build_benchmark_attempt_learning_gate,
+    build_benchmark_baseline_failure_gate_comparison,
+    build_benchmark_claim_review,
+    build_benchmark_learning_ledger,
+    build_benchmark_lifecycle_state,
+    build_benchmark_runner_invariant_review,
+    build_benchmark_verifier_attribution_review,
+)
+from .benchmark_adapters.agentissue import (
     AGENTISSUE_BENCHMARK_ID,
     AGENTISSUE_CODEX_CLI_RUNNER_EXECUTION_GATE_SCHEMA_VERSION,
     AGENTISSUE_CODEX_CLI_RUNNER_FIRST_RUN_HANDOFF_SCHEMA_VERSION,
@@ -42,69 +50,69 @@ from .benchmark import (
     materialize_agentissue_codex_cli_runner_synthetic_staging,
     materialize_agentissue_codex_cli_runner_target_handoff,
     materialize_agentissue_codex_cli_runner_workflow_check,
-    TERMINAL_BENCH_DEFAULT_DATASET,
-    TERMINAL_BENCH_DEFAULT_MODEL,
-    TERMINAL_BENCH_DEFAULT_TASK,
-    TERMINAL_BENCH_CODEX_INSTALL_STRATEGIES,
-    TERMINAL_BENCH_CODEX_INSTALL_STRATEGY_RUNTIME_INSTALL_IF_MISSING,
-    TERMINAL_BENCH_WORKER_CODEX_MATERIALIZATION_STRATEGIES,
-    TERMINAL_BENCH_WORKER_CODEX_MATERIALIZATION_STRATEGY_WORKER_PATH,
-    TERMINAL_BENCH_HARDENED_CODEX_BASELINE_PREFLIGHT_MODE,
-    TERMINAL_BENCH_MODES,
+)
+from .benchmark_adapters.agents_last_exam import (
+    AGENTS_LAST_EXAM_DEFAULT_ALT_DOCKER_IMAGE,
+    AGENTS_LAST_EXAM_DEFAULT_DOCKER_IMAGE,
+    AGENTS_LAST_EXAM_DEFAULT_SNAPSHOT,
+    build_agents_last_exam_baked_task_input_readiness,
+    build_agents_last_exam_baked_task_input_scan,
+    build_agents_last_exam_candidate_task_data_scan,
+    build_agents_last_exam_host_codex_cli_route,
+    build_agents_last_exam_host_codex_cua_no_task_smoke_from_environment,
+    build_agents_last_exam_local_dry_run_plan,
+    build_agents_last_exam_local_exact_dry_run_result,
+    build_agents_last_exam_local_launch_packet,
+    build_agents_last_exam_local_preflight,
+    build_agents_last_exam_local_runner_readiness,
+    build_agents_last_exam_local_source_readiness,
+    build_agents_last_exam_result_benchmark_report,
+    build_agents_last_exam_task_material_readiness,
+    build_agents_last_exam_validation_run_gate,
+)
+from .benchmark_adapters.skillsbench import (
     SKILLSBENCH_DEFAULT_DATASET,
     SKILLSBENCH_DEFAULT_MODEL,
     SKILLSBENCH_DEFAULT_ROUTE,
     SKILLSBENCH_DEFAULT_TASK,
     SKILLSBENCH_ROUTES,
-    build_agents_last_exam_result_benchmark_report,
-    TERMINAL_BENCH_MANAGED_CODEX_GOAL_HARNESS_KWARGS,
-    agent_kwargs_from_invocation,
-    build_agents_last_exam_local_dry_run_plan,
-    build_agents_last_exam_local_exact_dry_run_result,
-    build_agents_last_exam_host_codex_cua_no_task_smoke_from_environment,
-    build_agents_last_exam_host_codex_cli_route,
-    build_agents_last_exam_local_preflight,
-    build_agents_last_exam_local_runner_readiness,
-    build_agents_last_exam_local_source_readiness,
-    build_agents_last_exam_baked_task_input_readiness,
-    build_agents_last_exam_baked_task_input_scan,
-    build_agents_last_exam_candidate_task_data_scan,
-    build_agents_last_exam_task_material_readiness,
-    build_agents_last_exam_validation_run_gate,
-    build_agents_last_exam_local_launch_packet,
-    build_benchmark_claim_review,
-    build_benchmark_attempt_learning_gate,
-    build_benchmark_adapter_kwarg_absorption_review,
-    build_benchmark_baseline_failure_gate_comparison,
-    build_benchmark_candidate_source_boundary,
-    build_benchmark_learning_ledger,
-    build_benchmark_lifecycle_state,
-    build_benchmark_runner_invariant_review,
-    build_benchmark_verifier_attribution_review,
-    build_terminal_bench_result_finalization_gate,
-    build_terminal_bench_benchmark_run,
-    build_skillsbench_benchflow_result_benchmark_run,
     build_skillsbench_benchmark_run,
+    build_skillsbench_benchflow_result_benchmark_run,
+    skillsbench_recommended_action,
+)
+from .benchmark_adapters.terminal_bench import (
+    TERMINAL_BENCH_DEFAULT_DATASET,
+    TERMINAL_BENCH_DEFAULT_MODEL,
+    TERMINAL_BENCH_DEFAULT_TASK,
+    TERMINAL_BENCH_CODEX_INSTALL_STRATEGIES,
+    TERMINAL_BENCH_CODEX_INSTALL_STRATEGY_RUNTIME_INSTALL_IF_MISSING,
+    TERMINAL_BENCH_HARDENED_CODEX_BASELINE_PREFLIGHT_MODE,
+    TERMINAL_BENCH_MANAGED_CODEX_GOAL_HARNESS_KWARGS,
+    TERMINAL_BENCH_MODES,
+    agent_kwargs_from_invocation,
+    build_terminal_bench_benchmark_run,
     build_terminal_bench_environment_setup_probe_gate,
-    launch_terminal_bench_case_run,
     build_terminal_bench_harbor_result_benchmark_run,
+    build_terminal_bench_result_finalization_gate,
+    collect_terminal_bench_goal_harness_cli_bridge_trace,
     launch_terminal_bench_environment_setup_probe,
+    launch_terminal_bench_case_run,
     launch_terminal_bench_worker_materialization_probe,
     poll_terminal_bench_worker_materialization_probe,
     resume_terminal_bench_materialized_job,
-    benchmark_result_from_benchmark_run_for_baseline_gate,
-    collect_terminal_bench_goal_harness_cli_bridge_trace,
-    filter_public_benchmark_artifact_paths,
     summarize_terminal_bench_post_launch_materialization,
     terminal_bench_recommended_action,
-    skillsbench_recommended_action,
+    TERMINAL_BENCH_WORKER_CODEX_MATERIALIZATION_STRATEGIES,
+    TERMINAL_BENCH_WORKER_CODEX_MATERIALIZATION_STRATEGY_WORKER_PATH,
 )
 from .benchmark_ledger import (
     BENCHMARK_RUN_LEDGER_DEFAULT_PATH,
     update_benchmark_run_ledger,
 )
 from .benchmark_core import (
+    build_benchmark_candidate_source_boundary,
     build_codex_app_parity_posthoc_check,
+    filter_public_benchmark_artifact_paths,
     render_codex_app_parity_posthoc_check_markdown,
 )
 from .configure_goal import configure_goal, render_configure_goal_markdown
